Fix Incorrect Handling of Nested Objects
Operators $elemMatch and $in were not correctly handling nested objects.
For $elemMatch, "$and" was added to objects that hand more than one element.
This caused problems for the parser. Similarly, the $or operator inside
a nested $elemMatch caused an incorrect parsing. The $in operator was only
handling one object with one element and not multiple elements.

33294-query-text-search