Populate filtered rows
In some instances, the number of results returned by dreyfus
will be different than what we actually return to the user
due to our filtering match function. We fill in those gaps by
continously searching for rows until we can satisfy the limit. The
bookmark will reflect the last row which satisfies this condition.

33294-query-text-search